Nest function clasess into factory class
